A bunch of refactoring and work on binary tree stealers.
