+++ src/main/scala/scala/collection/par/workstealing/HashTries.scala
@@ -10,7 +10,7 @@ import scala.collection.immutable.HashMap
 import scala.collection.immutable.TrieIterator
 import scala.annotation.unchecked.{ uncheckedVariance => uV }
 
-object Trees {
+object HashTries {
 
   import Scheduler.{ Kernel, Node }
 
@@ -44,7 +44,7 @@ object Trees {
     def next(): T
   }
 
-  abstract class TrieChunkIterator[T, Repr] extends TrieIterator[T](null) with Trees.ChunkIterator[T] {
+  abstract class TrieChunkIterator[T, Repr] extends TrieIterator[T](null) with HashTries.ChunkIterator[T] {
     final def setDepth(d: Int) = unsafe.putInt(this, OFFSET_DEPTH, d)
     final def setPosD(p: Int) = unsafe.putInt(this, OFFSET_POS_D, p)
     final def setArrayD(a: Array[collection.immutable.Iterable[T]]) = unsafe.putObject(this, OFFSET_ARRAY_D, a)
@@ -75,12 +75,12 @@ object Trees {
   }
 
   trait Scope {
-    implicit def hashTrieSetOps[T](a: Par[HashSet[T]]) = new Trees.HashSetOps(a)
+    implicit def hashTrieSetOps[T](a: Par[HashSet[T]]) = new HashTries.HashSetOps(a)
     implicit def canMergeHashTrieSet[T: ClassTag](implicit ctx: Scheduler) = new CanMergeFrom[Par[HashSet[_]], T, Par[HashSet[T]]] {
       def apply(from: Par[HashSet[_]]) = new HashSetMerger[T](ctx)
       def apply() = new HashSetMerger[T](ctx)
     }
-    implicit def hashTrieMapOps[K, V](a: Par[HashMap[K, V]]) = new Trees.HashMapOps(a)
+    implicit def hashTrieMapOps[K, V](a: Par[HashMap[K, V]]) = new HashTries.HashMapOps(a)
     implicit def canMergeHashTrieMap[K: ClassTag, V: ClassTag](implicit ctx: Scheduler) = new CanMergeFrom[Par[HashMap[_, _]], (K, V), Par[HashMap[K, V]]] {
       def apply(from: Par[HashMap[_, _]]) = new HashMapMerger[K, V](ctx)
       def apply() = new HashMapMerger[K, V](ctx)
@@ -172,7 +172,7 @@ object Trees {
     val chunkIterator = new TrieChunkIterator[T, HashSet[T]] {
       final def getElem(x: AnyRef): T = {
         val hs1 = x.asInstanceOf[HashSet.HashSet1[T]]
-        Trees.key(hs1)
+        HashTries.key(hs1)
       }
     }
 
@@ -275,7 +275,7 @@ object Trees {
     val chunkIterator = new TrieChunkIterator[(K, V), HashMap[K, V]] {
       final def getElem(x: AnyRef): (K, V) = {
         val hm1 = x.asInstanceOf[HashMap.HashMap1[K, V]]
-        Trees.kv(hm1)
+        HashTries.kv(hm1)
       }
     }
     def next(): (K, V) = { nextProgress += 1; chunkIterator.next }
